fix/swagger in a separate file for sustainability
