+++ api/app.js
@@ -5,29 +5,13 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 var cors = require('cors');
 var swaggerJSDoc = require('swagger-jsdoc');
+
 const swaggerUi = require('swagger-ui-express');
-
-/*
-Configuration de Swagger
-*/
-const swaggerDefinition = {
-  openapi: '3.0.0',
-  info: {
-    title: 'API Trouve ton Artisan !',
-    version: '1.0.0',
-  },
-};
-
-const options = {
-  swaggerDefinition,
-  apis: ['./routes/*.js'],
-};
-
-const swaggerSpec = swaggerJSDoc(options);
+const swaggerSpec = require('./config/swagger');
+const sequelize = require('./config/database');
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
-const sequelize = require('./db/database');
 
 var app = express();
