+++ api/config/swagger.js
@@ -1,0 +1,24 @@
+const swaggerJsdoc = require('swagger-jsdoc');
+
+const options = {
+  definition: {
+    openapi: '3.0.0',
+    info: {
+      title: 'Trouve ton Artisan — API',
+      version: '1.0.0',
+      description: 'API REST de la plateforme artisans Auvergne-Rhône-Alpes',
+    },
+    servers: [
+      {
+        url: 'http://localhost:3001',
+        description: 'Serveur de développement',
+      },
+    ],
+  },
+
+  apis: ['./routes/*.js'],
+};
+
+const swaggerSpec = swaggerJsdoc(options);
+
+module.exports = swaggerSpec;
